cleaned up root directory
