+++ db/models/word.js
@@ -1,0 +1,4 @@
+const mongoose = require('mongoose');
+const schemas = require('../schemas');
+
+module.exports = mongoose.model('Word', schemas.wordSchema);
