+++ db/schemas/wordSchema.js
@@ -1,0 +1,5 @@
+const mongoose = require('mongoose');
+
+module.exports = mongoose.Schema({
+  word: { type: 'string', unique: true }
+});
